Moved: `modifier.py` to `src/pytorch/sparsification` Moved: training modifiers to 'pytorch/sparsification/training' Moved: distillation and activation_sparsity modifiers Moved: Respective tests
