+++ src/sparseml/pytorch/sparsification/pruning/modifier_pruning_base.py
@@ -27,7 +27,7 @@ from torch.optim.optimizer import Optimizer
 
 from sparseml.optim.modifier import BaseModifier
 from sparseml.pytorch.optim.analyzer_pruning import ModulePruningAnalyzer
-from sparseml.pytorch.optim.modifier import (
+from sparseml.pytorch.sparsification.modifier import (
     ModifierProp,
     ScheduledModifier,
     ScheduledUpdateModifier,
