+++ src/sparseml/pytorch/sparsification/training/__init__.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+
 # Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,22 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-Pruning modifiers and utilities to support their creation
-"""
-
-# flake8: noqa
-
-
-from .mask_creator import *
-from .mask_params import *
-from .modifier_as import *
-from .modifier_pruning_acdc import *
-from .modifier_pruning_base import *
-from .modifier_pruning_constant import *
-from .modifier_pruning_layer import *
-from .modifier_pruning_magnitude import *
-from .modifier_pruning_mfac import *
-from .modifier_pruning_movement import *
-from .modifier_pruning_structured import *
-from .scorer import *
+from .modifier_epoch import *
+from .modifier_lr import *
+from .modifier_params import *
+from .modifier_regularizer import *
