+++ tests/sparseml/pytorch/sparsification/distillation/test_modifier_distillation.py
@@ -21,9 +21,13 @@ from torch import Tensor
 from torch.nn import Module
 from torch.optim import Optimizer
 
-from sparseml.pytorch.optim import DistillationModifier, Modifier, ScheduledModifier
+from sparseml.pytorch.sparsification import (
+    DistillationModifier,
+    Modifier,
+    ScheduledModifier,
+)
 from tests.sparseml.pytorch.helpers import LinearNet, create_optim_sgd
-from tests.sparseml.pytorch.optim.test_modifier import ScheduledModifierTest
+from tests.sparseml.pytorch.sparsification.test_modifier import ScheduledModifierTest
 
 
 from tests.sparseml.pytorch.helpers import (  # noqa isort:skip
