+++ tests/sparseml/pytorch/sparsification/pruning/test_modifier_pruning_acdc.py
@@ -19,12 +19,12 @@ from torch.optim import SGD
 
 from sparseml.pytorch.sparsification.pruning import ACDCPruningModifier
 from tests.sparseml.pytorch.helpers import LinearNet
-from tests.sparseml.pytorch.optim.test_modifier import (
-    ScheduledModifierTest,
-    create_optim_adam,
-)
 from tests.sparseml.pytorch.sparsification.pruning.helpers import (
     state_dict_save_load_test,
+)
+from tests.sparseml.pytorch.sparsification.test_modifier import (
+    ScheduledModifierTest,
+    create_optim_adam,
 )
