+++ tests/sparseml/pytorch/sparsification/pruning/test_modifier_pruning_structured.py
@@ -24,14 +24,14 @@ from sparseml.pytorch.sparsification import (
 )
 from sparseml.pytorch.utils import tensor_sparsity
 from tests.sparseml.pytorch.helpers import LinearNet
-from tests.sparseml.pytorch.optim.test_modifier import (
-    ScheduledUpdateModifierTest,
-    create_optim_adam,
-    create_optim_sgd,
-)
 from tests.sparseml.pytorch.sparsification.pruning.helpers import (
     pruning_modifier_serialization_vals_test,
     sparsity_mask_creator_test,
+)
+from tests.sparseml.pytorch.sparsification.test_modifier import (
+    ScheduledUpdateModifierTest,
+    create_optim_adam,
+    create_optim_sgd,
 )
