+++ tests/sparseml/pytorch/sparsification/training/__init__.py
@@ -11,23 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-Pruning modifiers and utilities to support their creation
-"""
-
-# flake8: noqa
-
-
-from .mask_creator import *
-from .mask_params import *
-from .modifier_as import *
-from .modifier_pruning_acdc import *
-from .modifier_pruning_base import *
-from .modifier_pruning_constant import *
-from .modifier_pruning_layer import *
-from .modifier_pruning_magnitude import *
-from .modifier_pruning_mfac import *
-from .modifier_pruning_movement import *
-from .modifier_pruning_structured import *
-from .scorer import *
