+++ tests/sparseml/pytorch/sparsification/training/test_modifier_epoch.py
@@ -16,13 +16,13 @@ import os
 
 import pytest
 
-from sparseml.pytorch.optim import EpochRangeModifier
+from sparseml.pytorch.sparsification import EpochRangeModifier
 from tests.sparseml.pytorch.helpers import (
     LinearNet,
     create_optim_adam,
     create_optim_sgd,
 )
-from tests.sparseml.pytorch.optim.test_modifier import ScheduledModifierTest
+from tests.sparseml.pytorch.sparsification.test_modifier import ScheduledModifierTest
 
 
 from tests.sparseml.pytorch.helpers import (  # noqa isort:skip
